requirements.txt
discord.py
azure-identity
azure-mgmt-compute
flask